Cleaned Project1.m added functionality
Added scatter plot and cleaned the code a bit. It goes down to 10e-6 (500 iterations) for M=2 and M=4, for a training-length of 20, which good, and goes down to 10e-6 for a training-length of 150. We can compare M=4 and M=16 to see which one transmits more bits for part 2, that is if i dont get DFE working, which ill work on now. The code works for part 1 bpsk, by commenting out the convenc stuff and gets down to 10^-4 for training-length equal to 20.